map orthologous genes using ENSEMBL BioMART

findOrthologs(): map orthologous genes using ENSEMBL BioMART (getBM) from one species to another species. This function retrieves ids by 500 at a time to decrease the load on ENSEMBL servers.

findOrthologsHsMm(): map orthologous genes using ENSEMBL BioMART from human to mouse

findOrthologsMmHs(): map orthologous genes using ENSEMBL BioMART from mouse to human

loadBIOMARTdatasets(): Load 2 ENSEMBL BioMART datasets for finding orthologs. Details useMart

attributesFiltersFromTo(): List filters and attributes available in each dataset. Details: listAttributes

findOrthologs(datasets_FROM_TO = loadBIOMARTdatasets(from =
  "hsapiens_gene_ensembl", to = "mmusculus_gene_ensembl"),
  from_filters = "hgnc_symbol", from_values = c("TP53", "TERT"),
  to_attributes = "external_gene_name",
  to_homolog_attribute = "mmusculus_homolog_ensembl_gene",
  from_gene_id_name = "human_ensembl_gene_id",
  to_gene_id_name = "mouse_ensembl_gene_id")

findOrthologsHsMm(from_filters = "hgnc_symbol", from_values = c("TP53",
  "TERT"), to_attributes = "external_gene_name")

findOrthologsMmHs(from_filters = "ensembl_gene_id",
  from_values = c("ENSMUSG00000059552", "ENSMUSG00000021611"),
  to_attributes = "hgnc_symbol")

loadBIOMARTdatasets(from = "hsapiens_gene_ensembl",
  to = "mmusculus_gene_ensembl")

attributesFiltersFromTo(datasets_FROM_TO = loadBIOMARTdatasets())

Arguments

datasets_FROM_TO
environment containing "Mart" objects produced by loadBIOMARTdatasets(). It is convenient to specify datasets here rather than pre-loading. Full list of datasets in ENSEMBL BioMART: listDatasets("ensembl")

from_filters
Filters (one or more) that should be used in the query in the "from" species database. A possible list of filters can be retrieved using attributesFiltersFromTo(datasets_FROM_TO = loadBIOMARTdatasets()).

from_values
Values of the filter, e.g. vector of affy IDs in the "from" species database. If multiple filters are specified then the argument should be a list of vectors of which the position of each vector corresponds to the position of the filters in the filters argument. Please limit the number of values to less than 500.

to_attributes
Attributes you want to retrieve from the "to" species database. A possible list of attributes can be retrieved using the function attributesFiltersFromTo(datasets_FROM_TO = loadBIOMARTdatasets()).

to_homolog_attribute
Attribute that returns homolog gene id in the "from" database

from_gene_id_name
name of the column containing the ENSEMBL gene ids of "from" species

to_gene_id_name
name of the column containing the ENSEMBL gene ids of "from" species

from
ENSEMBL biomart dataset for a species whose identifiers you want to convert. Full list of possible options: listDatasets("ensembl")

to
ENSEMBL biomart dataset for a species whose identifiers you want convert into.

datasets_FROM_TO
environment containing "Mart" objects produced by loadBIOMARTdatasets()

Value

findOrthologs(): data.frame containing the mapping of homologous genes from one species to another species including to_attributes

findOrthologsHsMm(): data.frame containing the mapping of homologous genes from human to mouse including to_attributes

findOrthologsMmHs(): data.frame containing the mapping of homologous genes from mouse to human including to_attributes

loadBIOMARTdatasets(): environment containing 2 object of class 'Mart', one for each species dataset

attributesFiltersFromTo(): list containing 4 vector of either filters or attributes for each dataset

Examples

findOrthologs()

#>   human_ensembl_gene_id mouse_ensembl_gene_id external_gene_name hgnc_symbol
#> 1       ENSG00000141510    ENSMUSG00000059552              Trp53        TP53
#> 2       ENSG00000164362    ENSMUSG00000021611               Tert        TERT

findOrthologsHsMm()

#>   human_ensembl_gene_id mouse_ensembl_gene_id external_gene_name hgnc_symbol
#> 1       ENSG00000141510    ENSMUSG00000059552              Trp53        TP53
#> 2       ENSG00000164362    ENSMUSG00000021611               Tert        TERT

findOrthologsMmHs()

#>   mouse_ensembl_gene_id human_ensembl_gene_id hgnc_symbol
#> 1    ENSMUSG00000021611       ENSG00000164362        TERT
#> 2    ENSMUSG00000059552       ENSG00000141510        TP53

attributes = attributesFiltersFromTo(datasets_FROM_TO = loadBIOMARTdatasets())
grep("Mouse gene stable ID", attributes$from_attributes$description, ignore.case = T, value = T)

#> [1] "Algerian mouse gene stable ID"              
#> [2] "Mouse gene stable ID"                       
#> [3] "Northern American deer mouse gene stable ID"
#> [4] "Ryukyu mouse gene stable ID"                
#> [5] "Shrew mouse gene stable ID"                 
#> [6] "Steppe mouse gene stable ID"                

attributes$from_attributes[which(attributes$from_attributes$description == "Mouse gene name"),]

#>                                        name     description     page
#> 2213 mmusculus_homolog_associated_gene_name Mouse gene name homologs

grep("hgnc_symbol", attributes$from_attributes$name, value = T)

#> [1] "hgnc_symbol"

# find the desired column name
attributes$from_attributes[grep("Mouse gene stable ID", attributes$from_attributes$description, ignore.case = T),]

#>                                  name
#> 292     mspretus_homolog_ensembl_gene
#> 2212   mmusculus_homolog_ensembl_gene
#> 2308   pmbairdii_homolog_ensembl_gene
#> 2692     mcaroli_homolog_ensembl_gene
#> 2804     mpahari_homolog_ensembl_gene
#> 2900 mspicilegus_homolog_ensembl_gene
#>                                      description     page
#> 292                Algerian mouse gene stable ID homologs
#> 2212                        Mouse gene stable ID homologs
#> 2308 Northern American deer mouse gene stable ID homologs
#> 2692                 Ryukyu mouse gene stable ID homologs
#> 2804                  Shrew mouse gene stable ID homologs
#> 2900                 Steppe mouse gene stable ID homologs